IMPALA-4025: Part 1: Add percentile_disc aggregation function
This patch adds aggregation function percentile_disc. The implementation
is rewriting it into a subquery. The subquery computes the row number
on the ordering expr using analytic functions. The parent query then
picks the desired row using aggregation.
An Example of such rewrite is in StmtRewriter.java.

The behavior of this function is mostly the same as in PostgreSQL. The
handling of percentile expr not in [0, 1] is different: PostgreSQL
throws an error and impala returns NULL.

Some FE and EE tests are added. EE tests not related to the above
difference are verified against PostgreSQL.

Change-Id: Iacef7b3fcd74c4c73d88400ce27307c3baa0121e