IMPALA-5737: Tighten minicluster memory limit
This patch limits minicluster memory based on the memory available.
Two memory profiles are introduced. If there is less than 32GB memory
available, memory limit of the minicluster will be set to ~27GB.
Otherwise it will be set to ~32GB. This will hopefully help with
minicluster with more than 3 nodes.

The 27GB profile works on m2.4xlarge which has 8 cores. It should be
more than enough for 4-core machines.

Change-Id: If8240551e726c6da546a926a1ce3444f41ef87fe