IMPALA-7257. Support Kudu tables in LocalCatalog
This adds support for querying Kudu tables when --use_local_catalog is
enabled.

With this change, most of the kudu e2e tests pass. Those that don't are
related to missing support for CREATE TABLE AS SELECT, which currently
has some downcasts to 'Db' causing issues. That will be addressed in a
separate patch along with fixing CTAS for FS-backed tables.

Change-Id: I5b6a317ee895e43e00ade953e814867b56b4e6dd