IMPALA-7203. Support UDFs in LocalCatalog
This adds support to LocalCatalog to load persistent UDFs (both Java and
native) from the HMS. Transient UDFs are not supported, since, without a
central component to store them between restarts, there isn't any clear
path to doing so.

The various UDF tests do not yet pass reliably with this change since so
many of them rely on the transient Java UDF functionality. This is a
legacy feature that isn't commonly used today, and we may not be able to
support it in LocalCatalog.

For now, I tested manually that I could create, drop, and run some UDFs
and that they show up in 'show functions'.

Change-Id: I6130d07b9c641525382a618a9f8da048c7ae75ed